+++ pypinyin/style/_constants.py
@@ -88,9 +88,9 @@ RE_PHONETIC_SYMBOL = re.compile(
 )
 
 # 匹配使用数字标识声调的字符的正则表达式
-RE_TONE2 = re.compile(r'([aeoiuvnmê])([1-5])$')
+RE_TONE2 = re.compile(r'([aeoiuvnmêü])([1-5])$')
 # 匹配 TONE2 中标识韵母声调的正则表达式
-RE_TONE3 = re.compile(r'^([a-zê]+)([1-5])([a-zê]*)$')
+RE_TONE3 = re.compile(r'^([a-zêü]+)([1-5])([a-zêü]*)$')
 
 # 匹配单个数字
 RE_NUMBER = re.compile(r'\d')